# servicios/nucleo-financiero/src/main/resources/openapi/nucleo-financiero.yaml
...
openapi: 3.1.0
info:
  title: nucleo-financiero
  version: 1.0.0
  description: >-
    Servicio nucleo-financiero. Un caso de uso = una operacion, con el codigo CU en el
    operationId. Los importes viajan como CADENA decimal, nunca como number.
servers:
  - url: /api/v1
paths:
...
  /billetera/cotizaciones:
    post:
      operationId: cotizarOperacion            # CU-30 (consume tarifas)
      summary: Ver el precio de una recarga o un retiro antes de confirmarlo
      description: >-
        [CONTRATO] Que la persona vea el numero final ANTES de aceptar: base, comision, impuesto,
        costo total, neto y hasta cuando vale ese precio. No mueve nada ni compromete a nadie. La
        `Idempotency-Key` que se manda aca es la MISMA que despues lleva la recarga o el retiro, y
        `cotizacionId` es lo que esa operacion cita para confirmar. Pedir otro importe con la misma
        clave produce otra cotizacion: no se puede cobrar el precio de un importe sobre otro.

        El precio lo fija `tarifas`; este servicio no lo calcula. Una operacion sin concepto en el
        tarifario es gratuita (`gratuita=true`, sin `cotizacionId`).
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCotizacionOperacion' }
      responses:
        '200':
          description: La cotizacion y su desglose
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaCotizacionOperacion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-05 / AP-CU11-01 no se pudo cotizar (tarifas no respondio)
          # AP-CU10-08 / AP-CU11-12 el costo no puede igualar ni superar el importe
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/recargas:
    post:
      operationId: solicitarRecarga           # CU-10
      summary: Abre una orden de recarga y devuelve cuanto se va a acreditar
      description: >-
        [CONTRATO] Derivada de la implementacion del carril 2A, que llego antes que
        su contrato.

        La orden NO acredita: reserva la intencion y vence. Acreditar es el paso
        siguiente y ocurre cuando el proveedor confirma, porque acreditar contra una
        promesa es acreditar plata que puede no llegar.

        `acreditara` viaja en la respuesta ya calculado: el cliente **no** resta el
        costo del proveedor por su cuenta. Un cliente que recalcula produce un numero
        que no es el del libro contable.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRecarga' }
      responses:
        '201':
          description: Orden abierta, con su vencimiento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaRecarga' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-01 LIMITE_EXCEDIDO           · AP-CU10-02 SALDO_MAXIMO_ALCANZADO
          # AP-CU10-03 INSTRUMENTO_NO_VERIFICADO · AP-CU10-04 CUENTA_NO_OPERATIVA
          # AP-CU10-05 ORDEN_EXPIRADA
          # AP-CU10-06 COTIZACION_REQUERIDA · AP-CU10-07 COTIZACION_VENCIDA
          # AP-CU10-08 COTIZACION_NO_COINCIDE
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/recargas/{ordenId}/acreditacion:
    post:
      operationId: acreditarRecarga           # CU-10
      summary: Acredita una orden que el proveedor confirmo
      description: >-
        [CONTRATO] La confirmacion del proveedor. Es idempotente por la orden: si dos
        confirmaciones llegan juntas, una acredita y la otra encuentra la orden ya
        acreditada — la red duplica, y eso no puede duplicar el saldo.
      parameters:
        - name: ordenId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ClaveIdempotencia'
      responses:
        '200':
          description: Acreditada, con el saldo resultante
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaAcreditacion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-05 ORDEN_EXPIRADA
        '401':
          $ref: '#/components/responses/NoAutenticado'
...
components:
  parameters:
    ClaveIdempotencia:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Toda operacion con efecto la exige. La red duplica; reintentar tiene que ser
        seguro, y la clave se valida ANTES de escribir (invariante 7).
      schema: { type: string, format: uuid }
  schemas:
...
    EntradaCotizacionOperacion:
      type: object
      additionalProperties: false
      required: [operacion, cuentaBilleteraId, monto]
      properties:
        operacion: { type: string, enum: [RECARGA, RETIRO] }
        cuentaBilleteraId: { type: string, format: uuid }
        monto: { $ref: '#/components/schemas/Dinero' }
    SalidaCotizacionOperacion:
      type: object
      additionalProperties: false
      required: [base, comision, impuesto, costoTotal, neto, gratuita]
      properties:
        cotizacionId:
          type: string
          format: uuid
          description: la version del precio que se acepta; ausente cuando la operacion es gratuita
        base: { $ref: '#/components/schemas/Dinero' }
        comision: { $ref: '#/components/schemas/Dinero' }
        impuesto: { $ref: '#/components/schemas/Dinero' }
        costoTotal: { $ref: '#/components/schemas/Dinero' }
        neto:
          $ref: '#/components/schemas/Dinero'
        validaHasta: { type: string, format: date-time }
        gratuita: { type: boolean }
...
    EntradaRecarga:
      type: object
      additionalProperties: false
      required: [cuentaBilleteraId, monto, medio]
      properties:
        cuentaBilleteraId: { type: string, format: uuid }
        monto: { $ref: '#/components/schemas/Dinero' }
        costoProveedor: { $ref: '#/components/schemas/Dinero' }
        medio: { type: string, maxLength: 30 }
        instrumentoFondeoId: { type: string, format: uuid }
        cotizacionId:
          type: string
          format: uuid
          description: >-
            La cotizacion que la persona vio y acepto (POST /billetera/cotizaciones, con la MISMA
            Idempotency-Key). Obligatoria mientras `aportaya.cotizacion.exigir-previa` este activa.
    SalidaRecarga:
      type: object
      additionalProperties: false
      required: [ordenRecargaId, estado, expiraEn, acreditara]
      properties:
        ordenRecargaId: { type: string, format: uuid }
        estado: { type: string, maxLength: 20 }
        expiraEn: { type: string, format: date-time }
        acreditara:
          $ref: '#/components/schemas/Dinero'
    SalidaAcreditacion:
      type: object
      additionalProperties: false
      required: [ordenRecargaId, transaccionId, saldoDespues]
      properties:
        ordenRecargaId: { type: string, format: uuid }
        transaccionId:  { type: string, format: uuid }
        saldoDespues:   { $ref: '#/components/schemas/Dinero' }
...
    Dinero:
      type: object
      additionalProperties: false
      required: [monto, moneda]
      properties:
        monto:
          type: string
          pattern: '^-?\d+\.\d{2}$'     # CADENA: un number JSON es un doble
        moneda:
          type: string
          enum: [BOB, USD]
    Error:
      type: object
      additionalProperties: false
      required: [codigo, mensaje, trazaId]
      properties:
        codigo:   { type: string, pattern: '^AP-CU\d+-\d{2}$' }
        mensaje:  { type: string }
        detalle:  { type: object }
        trazaId:  { type: string }
...
  responses:
    ReglaDeNegocio:
      description: Regla de negocio de la aplicacion (422, no 400)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NoAutenticado:
      description: Sin sesion valida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }